// parinacoop-webserver-master/src/contexts/admin/application/create-client-use-case/create-client.dto.ts
export interface CreateClientDto {
  run: string;
  password: string;
  documentNumber: number;
  email: string;
  cellphone: string;
  names: string;
  firstLastName: string;
  secondLastName: string;
  typeAddress: string;
  street: string;
  number: number;
  detail: string;
  communeId: number;
}